requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
gspread
requests
beautifulsoup4
lxml
pandas
pdfplumber
pytesseract
pdf2image
Pillow
